feat: added graph and mermaid features to workflow object
